Relocate cursor operator tactics
