+++ differential-dataflow/src/operators/cursor/mod.rs
@@ -1,0 +1,4 @@
+//! Operator implementations based on cursor traversal.
+
+pub mod join;
+pub mod reduce;
